Move Popup components to subfolder
