+++ src/popup/components/Bar.vue
