+++ src/popup/components/ItemOverlay.vue
